Move static methods from loggers/base.py to utilities/loggers.py
