+++ tests/loggers/test_base.py
@@ -27,6 +27,7 @@ from pytorch_lightning.loggers.base import DummyExperiment, DummyLogger
 from pytorch_lightning.utilities import rank_zero_only
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.helpers.boring_model import BoringDataModule, BoringModel
+from pytorch_lightning.utilities.logger import _convert_params, _sanitize_params
 
 
 def test_logger_collection():
@@ -290,8 +291,8 @@ def test_np_sanitization():
 
         @rank_zero_only
         def log_hyperparams(self, params):
-            params = self._convert_params(params)
-            params = self._sanitize_params(params)
+            params = _convert_params(params)
+            params = _sanitize_params(params)
             self.logged_params = params
 
     logger = CustomParamsLogger()
